+++ lib/redgraph.rb
@@ -2,6 +2,8 @@
 require "redis"
 require "active_support/core_ext/hash/indifferent_access"
 require "active_support/core_ext/object/blank"
+require "active_support/core_ext/string/inflections"
+require "active_support/concern"
 
 require_relative "redgraph/version"
 require_relative "redgraph/util"
@@ -9,6 +11,7 @@ require_relative "redgraph/graph"
 require_relative "redgraph/node"
 require_relative "redgraph/edge"
 require_relative "redgraph/query_response"
+require_relative "redgraph/node_model"
 
 module Redgraph
   class Error < StandardError; end